+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Last Changes
 
+- [#24](https://github.com/LaxarJS/mashup-demo/issues/24): Updated tutorial for LaxarJS v2
 - [#21](https://github.com/LaxarJS/mashup-demo/issues/21): Fixed headline height
 - [#22](https://github.com/LaxarJS/mashup-demo/issues/22): Removed unnecessary alias in webpack configuration
 - [#20](https://github.com/LaxarJS/mashup-demo/issues/20): Upgraded to LaxarJS 2 and upgraded dependencies